requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.0.3
twilio==8.10.0
python-multipart==0.0.9
pydantic==2.4.2
google-cloud-speech==2.21.0
google-cloud-translate==3.12.0
google-cloud-vision==3.4.4
pandas==2.1.3
fpdf2==2.7.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pyzbar==0.1.9
Pillow==10.2.0
requests==2.31.0
SpeechRecognition==3.10.1
pydub==0.25.1
langdetect==1.0.9
nltk==3.8.1
aiohttp==3.9.3
opencv-python==4.9.0.80
numpy==1.26.4
ffmpeg-python==0.2.0 